[Common] Fixed postorder
Postorder did not include the entry block, i.e. block 0.
That was "convenient" because initially when it was
made, the entry block was not used in the fixed-point
iteration.

But this is bad, any algorithm that should skip the entry
block should be the one that actually skips it and not
the implementation of postorder.